Modularizes the functionality, making it simple to extend and configure - even outside of the library itself.  For example, one could write a custom module that extends BaseModule, and include that in the Quill module configuration alongside the built-ins without necessarily needing to submit a PR and get their extension included in the library.
